IDApi, WebAPI: an easier check for emptystrings
`typeof partnerParams` will always be "object" at this point, per just a few
lines above here, where we throw an error if it's anything different. That
means we probably have never really thrown this error for an empty string.

But even more to the point, if the value is missing or an empty string, then
the `typeof` and `length` checks will never run anyway, because both null and
emptystrings evaluate to `false`. So we can just drop the rest of the check.